fix: [AUDIT-5000-R3] BLUE SL priority, units float, GREEN TF description
Wave 3 audit (6 deep agents reading full transcriptions) found 8 issues:

1. BLUE SL: Prioritize previous minimum over Fib 0.618 (HIGH).
   Alex: "siempre protegemos el minimo anterior, 0.618 es orientacion"
   Added 0.2% buffer ("poquito de espacio") below SL level.
2. units: int → float in TradeRisk and ManagedPosition to support
   fractional crypto lots (0.001 BTC). DB was already REAL.
   Position sizing now preserves decimals for small units (<100).
3. GREEN AI prompt: Fixed incorrect claim that GREEN has "FIXED
   timeframes regardless of style". Mentorship teaches 3 different
   TF sets (swing: W→D→1H→15M, day: 4H→1H→15M→2M, scalp: 15M→5M→1M→30s).

731/731 tests passing. Runtime verified.

+++ backend/core/position_manager.py
@@ -154,7 +154,7 @@ class ManagedPosition:
     current_sl: float
     take_profit_1: float
     take_profit_max: Optional[float] = None
-    units: int = 0                  # Position size (signed: +BUY, -SELL)
+    units: float = 0.0              # Position size (signed: +BUY, -SELL), float for crypto fractional lots
     style: str = "day_trading"     # Trading style: "day_trading", "swing", "scalping"
     strategy_variant: Optional[str] = None  # e.g. "GREEN", "BLUE_A", "RED" — for strategy-specific logic
     trailing_tp_only: bool = False  # True for crypto GREEN: skip hard TP1 close, use EMA 50 trailing
